+++ src/app.js
@@ -1,15 +1,8 @@
 import React from 'react'
+import Router from 'react-router'
 import styles from './app.less'
+import routes from './routes.js'
 
-class Hello extends React.Component{
-  render () {
-    return <div className='container'>
-        <header role='banner'>
-          <h1 style={{fontWeight: '100', fontSize: '4rem', marginTop: '16rem'}}>boilerplate-hjs-webpack</h1>
-        </header>
-        <p>Check it out on <a href={'https://github.com/clintonhalpin/boilerplate-hjs-webpack'} style={{ color: 'white'}}>github</a> <br /><br/> built by <a href={'http://twitter.com/clintonhalpin'} style={{ color: 'white'}}>@clintonhalpin</a><br /><br />taken from <a href={'https://github.com/HenrikJoreteg/hjs-webpack/tree/master/examples/assets-and-index-html'} style={{ color: 'white'}}>hjs-webpack</a></p>
-      </div>
-  }
-}
-
-React.render(<Hello/>, document.body)
+Router.run(routes, (Root, state) => {
+  React.render(<Root {...state} />, document.body);
+});
