+++ src/components/Hello2.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+class Hello2 extends React.Component{
+  render () {
+    return (
+        <div className='container'>
+        <header role='banner'>
+          <h1 style={{fontWeight: '100', fontSize: '4rem', marginTop: '16rem'}}>Another Route!</h1>
+        </header>
+      </div>
+    )
+  }
+}
+
+export default Hello2;
+
+
